pyproject.toml
[tool.poetry]
name = "pr-report"
version = "0.1.0"
description = ""
authors = ["Andrew Sosa <andrew@andrewsosa.dev>"]
packages = [
    { include = "pr_report", from = "src" },
]

[tool.poetry.scripts]
pr-report = "pr_report.cli:cli"

[tool.poetry.dependencies]
python = "^3.7"
PyGithub = "^1.54.1"
tqdm = "^4.57.0"
requests-cache = "^0.5.2"
huepy = "^1.2.1"
click = "^7.1.2"
logzero = "^1.6.3"

[tool.poetry.dev-dependencies]
flake8 = "^3.8.4"
isort = "^5.7.0"
black = "^20.8b1"
mypy = "^0.800"

[tool.black]
line_length = 88

[tool.isort]
profile = 'black'
line_length = 88
float_to_top = true
order_by_type = false
use_parentheses = true

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
